Remove broken Heroku support
This attempt failed. For json-server to serve the API in a way that is
accessible to the application, it would need to be a "web" process. You
can only have one of those per Heroku application.

json-server might be able to serve the API correctly on Heroku if it
were its own application, but that would be its own form of headache.