Do show chart animations for overlay
I disabled animations in the overlay because I thought it was overkill.
If a user chooses to make a chart fullscreen, they don't need to see the
animation all over again.

But the animation setting also applies to the animation that is run when
a population is enabled or disabled. There doesn't appear to be a way to
animate one of these events but not the other, so I'd rather animate
them both.