Fix #13: Add favicon
This icon is a modification of the following icon from Icons8:
https://icons8.com/icon/19729/thin-test-tube

Their icons are free without attribution for popular open-source
projects, which they define as projects with a community of users and
contributors:
https://icons8.com/license
https://icons8.com/articles/weve-contributed-an-icon-to-imageoptim/

Their definition of "popular" isn't perfectly precise, but I would hope
that any open-source project at Mozilla would qualify. And this is a
modification of an icon as it is.