Move front-end to repository root
As an obsessively organized person this sort of kills me. But there's no
other way around it other than splitting the front-end into its own
repository, which we decided not to do.

Greenkeeper and create-react-app-buildpack both insist that package.json
be at the root of the repository. react-scripts insists that
package.json be in the same directory as the src and public directories.
(And it's not possible to rename them, either.)

So we're stuck with weirdly-named src and public folders at the root of
the repository. Oh well.