Fix #1: Paginate the tables on the homepage
I decided to remove sorting as part of this. I had it working at one
point, but the user experience of sorting while also paginating was
really weird. I imagine most people would have been confused by it.

I don't think sorting is as useful when pagination is possible and
experiments are reverse-sorted by start date. And if there are cases
where someone really wants to find an experiment by name, search would
probably be a better way of doing that.