Turn off chart animations
Chart animations can be distracting. They are triggered whenever a chart
is rendered, meaning that animations happen whenever charts are updated,
whenever a user performs a search, and more.

The animations that play when the populations are changed can be useful,
but unfortunately there doesn't appear to be a way to enable *only*
those animations.