Add "npm audit" after tests, fix errors
This commit also switches the project from Yarn to NPM. NPM has
lockfiles now, which was one of the primary benefits of using Yarn, and
Yarn does not currently have an "audit" command. It plans to add one,
but even then, it's unclear that it will add an "npm audit fix" command
which is incredibly useful.